requirements.txt
Django==2.1.5
djangorestframework==3.9.1
python-decouple==3.1
pytz==2018.9
psycopg2-binary==2.7.7
